# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResADMIN": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAdmin": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqADMIN": {
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqADMIN": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IADMIN": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteAdmin": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminListReqPRODUCT": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResPRODUCT": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"VendorId": {
						"type": "number",
						"format": "double"
					},
					"PizzaName": {
						"type": "string"
					},
					"PizzaDescrpation": {
						"type": "string"
					},
					"PizzaPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"VendorId",
					"PizzaName",
					"PizzaDescrpation",
					"PizzaPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqPRODUCT": {
				"properties": {
					"VendorId": {
						"type": "number",
						"format": "double"
					},
					"PizzaName": {
						"type": "string"
					},
					"PizzaDescrpation": {
						"type": "string"
					},
					"PizzaPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"VendorId",
					"PizzaName",
					"PizzaDescrpation",
					"PizzaPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqPRODUCT": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"VendorId": {
						"type": "number",
						"format": "double"
					},
					"PizzaName": {
						"type": "string"
					},
					"PizzaDescrpation": {
						"type": "string"
					},
					"PizzaPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"VendorId",
					"PizzaName",
					"PizzaDescrpation",
					"PizzaPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPRODUCT": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteProduct": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductListReqPRODUCT": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"VendorId": {
						"type": "number",
						"format": "double"
					},
					"PizzaName": {
						"type": "string"
					},
					"PizzaDescrpation": {
						"type": "string"
					},
					"PizzaPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"VendorId",
					"PizzaName",
					"PizzaDescrpation",
					"PizzaPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResORDER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"orderQty": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"orderStatus": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"orderAmount": {
						"type": "number",
						"format": "double"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"orderQty",
					"productId",
					"orderStatus",
					"address",
					"orderAmount",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetORDER": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqORDER": {
				"properties": {
					"orderQty": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"orderStatus": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"orderAmount": {
						"type": "number",
						"format": "double"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"orderQty",
					"productId",
					"orderStatus",
					"address",
					"orderAmount",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqORDER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"orderQty": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"orderStatus": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"orderAmount": {
						"type": "number",
						"format": "double"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"orderQty",
					"productId",
					"orderStatus",
					"address",
					"orderAmount",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IORDER": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteORDER": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderListReqORDER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"orderQty": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "string"
					},
					"orderStatus": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"orderAmount": {
						"type": "number",
						"format": "double"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"orderQty",
					"productId",
					"orderStatus",
					"address",
					"orderAmount",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "yummipizzaapp",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/getAdmin": {
			"post": {
				"operationId": "Getadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResADMIN"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/saveadmin": {
			"post": {
				"operationId": "Saveadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResADMIN"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqADMIN"
							}
						}
					}
				}
			}
		},
		"/admin/updateadmin": {
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResADMIN"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqADMIN"
							}
						}
					}
				}
			}
		},
		"/admin/deleteadmin": {
			"delete": {
				"operationId": "Deletadmin",
				"responses": {
					"200": {
						"description": "Admin deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IADMIN"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/getadminlist": {
			"post": {
				"operationId": "GetadminList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AdminListReqPRODUCT"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": []
			}
		},
		"/product/getProduct": {
			"post": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResPRODUCT"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaItems"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetProduct"
							}
						}
					}
				}
			}
		},
		"/product/saveProduct": {
			"post": {
				"operationId": "SaveProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResPRODUCT"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaItems"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqPRODUCT"
							}
						}
					}
				}
			}
		},
		"/product/updateProduct": {
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResPRODUCT"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaItems"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqPRODUCT"
							}
						}
					}
				}
			}
		},
		"/product/deleteProduct": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Product deleted Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPRODUCT"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaItems"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteProduct"
							}
						}
					}
				}
			}
		},
		"/product/getProductList": {
			"get": {
				"operationId": "GetProductList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductListReqPRODUCT"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaItems"
				],
				"security": [],
				"parameters": []
			}
		},
		"/order/getOrder": {
			"post": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResORDER"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaOrders"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetORDER"
							}
						}
					}
				}
			}
		},
		"/order/saveOrder": {
			"post": {
				"operationId": "SaveOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResORDER"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaOrders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqORDER"
							}
						}
					}
				}
			}
		},
		"/order/updateOrder": {
			"put": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResORDER"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaOrders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqORDER"
							}
						}
					}
				}
			}
		},
		"/order/deleteOrder": {
			"delete": {
				"operationId": "DeleteOrder",
				"responses": {
					"200": {
						"description": "Order deleted Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IORDER"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaOrders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteORDER"
							}
						}
					}
				}
			}
		},
		"/order/getOrderList": {
			"get": {
				"operationId": "GetOrdersList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrderListReqORDER"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"PizzaOrders"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}